server/requirements.txt
Django==3.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework==3.11.1
Pillow==7.2.0
dj-database-url==0.5.0
djangorestframework-simplejwt==4.4.0
ipython==7.17.0

# dev dependencies
django-debug-toolbar==2.2
